mmc: dw_mmc: Cleanup disable of low power mode w/ SDIO interrupts
In the patch (9623b5b mmc: dw_mmc: Disable low power mode if SDIO
interrupts are used) I added code that disabled the low power mode of
dw_mmc when SDIO interrupts are used.  That code worked but always
felt a little hacky because we ended up disabling low power as a side
effect of the first enable_sdio_irq() call.  That wouldn't be so bad
except that disabling low power involves a complicated process of
writing to the CMD/CMDARG registers and that extra process makes it
difficult to cleanly the read-modify-write race in
dw_mci_enable_sdio_irq() (see future patch in the series).

Change the code to take advantage of the init_card() callback of the
mmc core to know when an SDIO card has been inserted.  If we've got a
SDIO card and we're configured to use SDIO Interrupts then turn off
"low power mode" right away.

BUG=chrome-os-partner:22915
TEST=Boot and can use WiFi; suspend/resume is OK.

Conflicts (picking kernel-next version):
	drivers/mmc/host/dw_mmc.c
	include/linux/mmc/dw_mmc.h
...one conflict was there because of a slightly different way that the
original SDIO interrupt code landed in 3.4 (compared with upstream).

Change-Id: I8f343ad58f4b7d396da619eb62c4c91cea6d5c7a
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173131
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Ashok Nagarajan <asnagarajan@chromium.org>