CHROMIUM: dts: mmc: update device tree for upstream handling of SMU
Upstream chose to handle the SMU (Security Management Unit) quirk by
adding a distinct type of dwmmc controller.  Since we've picked up the
upstream patch, we need to update our device tree for 5420 to reflect
that.  Without this wifi fails to initialize properly and we timeout
trying to load the firmware.  This patch also corrects the reg property
on the non-SMU capable instance of dwmmc.

BUG=chrome-os-partner:20212
TEST=peach-pit boots and wifi comes up properly

Change-Id: I110a2caad758d18a9fc90602d29813c1f7f80057
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173522