ARM: dts: Apply INT rail workaround to all variants of exynos5420-peach-pit
Previously we were told that the INT rail workaround would not be
needed going forward for newer revisions of peach-pit.  It now appears
that's not the case.  Move the workaround to the common peach-pit dts
file and remove it from the revisions.

While we haven't measured the exact power hit of keeping the INT rail
on (in low power mode) during suspend, we have done measurements of S3
power on rev5 with the INT rail workaround in place and found it to be
acceptible.

BUG=chrome-os-partner:22082
TEST=hexdump -C /proc/device-tree/*/*/*/BUCK3/*op-mode => 00 00 00 02
TEST=Suspend/resume on all revisions still good.

Change-Id: I3112e6ce8102e2ae8811ab1045a319d680cf817b
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173210
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>