mmc: dw_mmc: Support setting the ciu-div on exynos
The existing dw_mmc code for exynos5250/exynos5420 read the ciu-div
property from the device tree and used it in calculations, but didn't
actually ever set it in hardware.  That means that if the hardware
value set by the firmware didn't match what was in the device tree
that we'd be clocking things incorrectly.

BUG=chrome-os-partner:23260
TEST=Try setting the ciu-div to something else in the device tree file
and validate that things work.

Change-Id: Icf86f16d6afb74608ec635eef9174fae7a81251a
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173233
Reviewed-by: Grant Grundler <grundler@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>