CHROMIUM: dts: peach-pit add vqmmc regulator for dwmmc2 SD cards
This wires up the regulator to the device node for the mmc controller
that's connected to the SD cards.  This will let us do
voltage-switching for the I/O properly and get UHS speeds.
It also amends the min-volatage for LDO4 on the 77802 PMIC to let us
go down to 1.8v.

BUG=chrome-os-partner:20212
TEST=none (needs other patches in series)

Change-Id: I9b77ac8007410662a69baa218342f3390bec0f07
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173532