drm/exynos: replace hdmi reset with hdmi disable
Before setting the core and timing generation registers, hdmi driver
resets the whole hdmi hardware, which also resets the audio related
registers. Hdmi Audio driver is not aware of this reset and doesn't
initialise the audio related registers back to the previous values.

Hdmi reset is replaced by hdmi disable which is called just before
setting the core and timing registers. It also ensure that audio
settings are not changed.

BUG=chrome-os-partner:22200
TEST=System works fine on PEACH and SNOW board. Tested dock mode
<--> normal mode along with audio playback.

Change-Id: I96a5fe456d2efec1ac89d563134037b9e69f0ce5
Signed-off-by: Rahul Sharma <rahul.sharma@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/173683
Reviewed-by: Akshu Agrawal <akshu@chromium.org>