CHROMIUM: s5p-mfc: Don't try to resubmit VP8 bitstream buffer for decode.
Currently, for formats that are not H264 (which means VP8 for CrOS), MFC
driver will check the consumed stream size returned by the firmware and,
based on that, will try to decide whether the bitstream buffer contained
more than one frame. If the size of the buffer is larger than the consumed
stream, it assumes that there are more frames in the buffer and that the
buffer should be resubmitted for decode. This rarely works though and
actually introduces problems (like crbug.com/308816), because:

- v7 firmware will always return consumed stream size equal to whatever
the driver passed to it when running decode (which is the size of the whole
buffer), which means we will never try to resubmit, because the firmware
will always tell us that it consumed all the data we passed to it;

- v6 firmware will return the number of consumed bytes, but will not
include the padding ("stuffing") bytes that are allowed after the frame
in VP8. Since there is no way of figuring out how many of those bytes
follow the frame without getting the frame size from IVF headers (or
somewhere else, but not from the stream itself), the driver tries to guess that
padding size is not larger than 4 bytes, which is not always true;

The only way to make it work is to queue only one frame per buffer from
userspace and the check in the kernel is useless and wrong for VP8.
MPEG4 still seems to require it, so keep it only for that format.

BUG=chromium:308816
TEST=VP8 playback on 3.8 ARM platforms and webrtc HW decode

Change-Id: I93e69f8f7d2f1dd71122fde56a2ba1089f153269
Signed-off-by: Pawel Osciak <posciak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173854
Reviewed-by: Kiran AVND <kiran@chromium.org>