ARM: dts: Apply "edge-triggered-interrupt" flag to daisy and peach-pit
This flag indicates to dw_mmc and ultimately the MMC core that
the controller struggles to deal with multiple back-to-back
interrupts.  This is caused at least in part by the dw_mmc working
in an edge-triggered fashion while the SDIO bus assumes the
the controller works as level-triggered.

BUG=chrome-os-partner:22915
TEST=During sustained simultaneous WiFi / Bluetooth activity, watch for
command timeouts.
TEST=suspend_stress_test

[dianders: re-picking original CL now that (6e7f078 CHROMIUM: sdio:
Defer SDIO interrupt handling until after resume) has landed]

Change-Id: I2f36c0e528d4aecf3678bc1ec96279f53bade8de
Signed-off-by: Paul Stewart <pstew@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previously-Reviewed-on: https://chromium-review.googlesource.com/174242
Reviewed-on: https://chromium-review.googlesource.com/175606