ARM: dts: Apply "edge-triggered-interrupt" flag to daisy and peach-pit
This flag indicates to dw_mmc and ultimately the MMC core that
the controller struggles to deal with multiple back-to-back
interrupts.  This is caused at least in part by the dw_mmc working
in an edge-triggered fashion while the SDIO bus assumes the
the controller works as level-triggered.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:22915
TEST=During sustained simultaneous WiFi / Bluetooth activity, watch for
command timeouts.

Change-Id: I0ff4581267f1ceecc7f3a52cc2acbbfe75b5c4dc
Reviewed-on: https://chromium-review.googlesource.com/174242
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>