CHROMIUM: atmel_mxt_ts: optionally start device on lid event
The lid open event comes very early in the resume sequence. So
early, in fact, that the underlying I2C buses may not have been
resumed yet. This can lead to deadlocks. Therfore, if the device
is coming out of suspend, don't start the device on lid event.
Instead, start the device as part of the resume callback.

BUG=chrome-os-partner:23475
TEST=1. suspend/resume the system using lid close/open as well as idle
suspend and check that the touch screen continues to function upon
resuming.
2. Stress test the lid open/close suspend/resume case either by waving a
magnet over the lid sensor or servo and check that no kernel panics
happen.

Change-Id: I07575596af0011b1857437cf0e5984b66a10ec97
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/174904
Reviewed-by: Benson Leung <bleung@chromium.org>