CHROMIUM: dw_mmc: Signal edge-triggered interrupts
Add a flag to the device tree that indicates that SDIO interrupts
are unreliable, i.e., they don't act like level-triggered interrupts.
This signals to the SDIO core that it should perform a hybrid of
interrupt-triggered and polling interrupts.

This is implemented by performing an agressive exponential back-off
in polling.  When the the system becomes idle this reverts to an
interrupt-only mode.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:22915
TEST=During sustained simultaneous WiFi / Bluetooth activity, watch for
command timeouts.

Change-Id: I968222367542483e873352ee1161d833507dd741
Reviewed-on: https://chromium-review.googlesource.com/175345
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>