drm/i915: Fix fbc + rc6 combination on SNB
On Sandybridge we must set the "PPGTT Render Target Base Address Valid
for FBC" bit as noted in the programming guide. We did this at clock
gating init. Thisbit is not saved and restored with RC6 power context,
so the resetting it at ring flush should fix that.

The effect of not doing this should be corruption, and not a hang - as
has so often been the case.

Note that we should actually clear this bit as well when not blitting to
the scanout (using the blitter for other things), or else all operations

Cc: Stéphane Marchesin <marcheu@chromium.org>
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>

BUG=chromium:303986
TEST=by hand:
i=0; while true; do i=$(($i+1)); echo "*** iteration $i"; X & sleep 2; DISPLAY=:0 xrandr --auto ; sleep 1; killall X; sleep 2; done

Change-Id: Ic0433466c2e466c7e0e934efc03d3f67b9c977d4
Reviewed-on: https://chromium-review.googlesource.com/175541
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>