MALI: remove redundant new lines from KBASE_DEBUG_*() format strings
KBASE_DEBUG_*() macros always append a '\n' at the end of the format
string.  If the format string ends in '\n', this just leads to an extra
blank line in dmesg.

There is a depressing number of checkpatch problems (4 errors, 47 warnings)
in this patch, but I'll leave it to arm to fix these in the next release.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:20550
TEST=builds clean, no extra newlines after Mali messages in
  /var/log/messages

Change-Id: Ief287048bc9afd3d15e2973702d148d6f56b7f5d
Reviewed-on: https://chromium-review.googlesource.com/174707
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>