drm/exynos: Avoid disabling HDMI TG registers
According to the SLSI hardware team, disabling the TG registers in the
HDMI block clears the mixer registers. This results in undefined
behavior in the mixer such as the inability to suspend/resume and
a blank screen on dpms on/off.

This patch removes the TG disable from the power off routine.

BUG=chrome-os-partner:22200,chrome-os-partner:23948
TEST=Verified S2R for audio and hotplug.
Below script now works after this patch:
	X -nocursor &
	DISPLAY=:0.0 xrandr --output eDP-1 --mode 1366x768 --pos
		0x0 --output HDMI-1 --mode 1024x768 --pos 1366x0
        DISPLAY=:0.0 xrandr --output eDP-1 --mode
		1366x768 --output HDMI-1 --mode 1024x768
		--same-as eDP-1
	DISPLAY=:0.0 display -geometry
		1024x768+0+0
		/usr/bin/colorbar.png

Change-Id: Ib04e7725d7a02c3c029cc2f63657b230f5b8064e
Signed-off-by: Shirish S <s.shirish@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/175913
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>