arm: exynos5250: Save and restore ACTLR across suspend/resume
The contents of the ACTLR is lost on kernel 3.8 across a
suspend/resume cycle. This means that the workarounds for errata
773022 and 774769 are lost. 773022 in particular is definitely
triggered by webkit and causes branch instructions to fail.

Thankfully this only affects 5250 which has an earlier A15
silicon than 5420, so let's make it specific to that.

BUG=chromium:308864
TEST=by hand: suspend/resume a daisy, go to amazon, webkit doesn't crash

Change-Id: I58c3b6fd6c50cf26ae12a7bae831cf5c990fccf1
Reviewed-on: https://chromium-review.googlesource.com/176031
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>