CHROMIUM: mmc: dw_mmc: Fix the max_blk_count in IDMAC
max_blk_count is currently set incorrectly, and the correct value can
be determined by looking at the maximum number of bytes which can be
transferred and the block size.  We use the maximum block size to
determine the minimum number of blocks that the controller should
support.  It could probably do more with a smaller block size but this
value should be sufficient for good performance.

This improves sequential performance by 82% on write and 6% read on
one particular device.

BUG=chrome-os-partner:24010
TEST=build and booted on pit
TEST= with this patch
	time dd if=/dev/zero of=/usr/local/100 bs=1M count=100 conv=fdatasync
	100+0 records in
	100+0 records out
	104857600 bytes (105 MB) copied, 2.62541 s, 39.9 MB/s

	real    0m2.638s
	user    0m0.000s
	sys     0m0.645s

	without this patch:
	time dd if=/dev/zero of=/usr/local/100 bs=1M count=100 conv=fdatasync
	100+0 records in
	100+0 records out
	104857600 bytes (105 MB) copied, 3.25873 s, 32.2 MB/s

	real    0m3.265s
	user    0m0.005s
	sys     0m0.690s

Change-Id: Ib3e2abe0ebbf0d0d838433d16e06387f11ecc73b
Signed-off-by: Seungwon Jeon <tgih.jun@samsung.com>
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/176488