CHROMIUM: ARM: make stack dump format consistent
Before this change, the function dump_backtrace() prints "Backtrace: "
at the beginning of a stack dump when CONFIG_ARM_UNWIND is n, but
does not print it when CONFIG_ARM_UNWIND is y.  This makes it much
harder for the crash reporter to find the beginning of the stack.

BUG=317788
TEST=not tested
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: If18bb3d7f51fcecb5aca4c6e1c6ccc5a9c834d2a
Reviewed-on: https://chromium-review.googlesource.com/176479
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>