UPSTREAM: NFC: Return driver failure upon unknown event reception
If the device is polling, this will trigger a netlink event to notify
userspace about the polling error.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>
(cherry picked from commit 500d824bdfac5086825d200d9a3bc5125043d4d4)
https://git.kernel.org/cgit/linux/kernel/git/sameo/nfc-next.git/commit/?id=500d824bdfac5086825d200d9a3bc5125043d4d4

BUG=chromium:320894
TEST=Reproduce the particular error case described by crbug.com/320894
by using a Nexus 4 as a p2p device and a PN544C3 as the HCI based NFC
adapter. Verify that a netlink event is generated for the failure and
that neard sees it.

Change-Id: I10eb2425dba8a082490300b8f1de1aafa735a87c
Signed-off-by: Arman Uguray <armansito@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/177273
Reviewed-by: Benson Leung <bleung@chromium.org>