drm/anx7808: Add HDCP support in anx7808 driver
This patch allows HDCP to be turned on/off in anx7808 based on
a drm connector property. When the HDCP property is set to desired,
HDCP is first enabled for [AP HDMI Tx --> 7808 HDMI Rx]. The
anx7808 driver waits for an interrupt that signals the start of
upstream HDCP authentication, then HDCP is enabled for
[7808 DP Tx --> 7730 dongle or monitor DP Rx]. When the HDCP
property is set to off, HDCP is first disabled between AP
and 7808, then disabled for 7808's downstream.

anx7808_reset_hdcp() resets 7808 DP Tx HDCP when a downstream
error is detected. Downstream authentication is retried up to
10 times before giving up. Note [AP HDMI Tx --> 7808 HDMI Rx]
HDCP link remains intact all the time.

If the monitor/downstream does not support HDCP, or downstream
always fails to authenticate, the HDCP property should stay as "Desired".

When 7808's downstream HDCP authentication completes successfully,
the HDCP property is updated to "Enabled".

BUG=chrome-os-partner:23567
TEST=Tested on spring

Change-Id: I5ae87ea61c356e69fd3a5451743a7ad06ad3a346
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/174547
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>