MALI: Separate PM callbacks for suspend/resume
The existing platform-specific API exposed callbacks to turn the GPU
on and off. These were used by the power-policy (on idle), and on system
suspend.

However, it is likely a platform needs to perform different operations
on suspend. In the case of ChromiumOS on Exynos we need to reset the
DVFS before suspending. We also need to ensure we always reset the GPU
on resume, as we don't reset the GPU after idle (as the GPU is not
powered off).

BUG=chromium:316186
BUG=chrome-os-partner:23950
TEST=build and run on daisy; ensure the system suspends correctly

Signed-off-by: David Garbett <david.garbett@arm.com>
Change-Id: Ibbac3d3db09b3ed292bd9d06e915e46055666796
Reviewed-on: https://chromium-review.googlesource.com/176830
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>