drm/exynos: do not copy adjusted mode into mode during crtc mode_set
The 'mode' is the modeline information which specifies the ideal mode
requested by the mode set initiator (usually userspace).
The 'adjusted_mode' is the actual hardware mode after all the crtcs
and encoders have had a chance to "fix it up".

The adjusted_mode starts as a duplicate of the mode in
drm_crtc_helper_set_mode(), and gets modified as required.  There is no
reason to touch the original requested mode.

In fact, doing so will cause us to think a new mode is being requested
whenever userspace tries to establish a new framebuffer with
drmModeSetCrtc(), since userspace will still be using the ideal mode, but
the crtc will be incorrectly comparing it against the adjusted_mode.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:266008
TEST=manual on daisy, daisy_spring, peach_pit: verify only 1 mode set at
boot (no black flashing) at splash screen or login prompt.

Change-Id: I04e549f9f8ebb289ae1cceb3f6e9bb8f1487b79d
Reviewed-on: https://chromium-review.googlesource.com/177017
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>