Revert "drm/exynos: Don't adjust the clock in mode_fixup"
This reverts commit a5bb726d0b8149fc0d0678bf255bcd6e4943b641.

The root cause of the many modesets at boot was that we were
overwriting the crtc's mode with the adjusted (hardware) mode.
If we don't overwrite crtc->mode, it will agree with the mode
specified by userspace and won't trigger a mode set.
The adjusted_mode gets cached in crtc->hwmode, so it can be
used by drm to compute acurate pixel times, etc.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:266008
TEST=Test on daisy, daisy_spring, peach_pit.
  => No black flashing during splash screen/login screen
  => Only one LCD modeset at boot

Change-Id: I76752e5809090b1c596476be3555d385f54cf31c
Reviewed-on: https://chromium-review.googlesource.com/177018
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>