drm/exynos: Reject invalid hdmi modes in mode_fixup
This patch moves the mode rejection logic in the hdmi and mixer drivers
from the get_modes connector path into the modeset path. The advantage
to doing this is that we no longer need to tie mixer to hdmi and assume
hdmi version x maps to mixer version y. Another benefit is that it
allows us to use connectors implemented in third party bridges.

BUG=chromium:248816
TEST=Tested on spring, choose 1280x1024 & ensure it's rejected

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Ie10eb9514536e31a833979a700b1fc3c43522416
Reviewed-on: https://chromium-review.googlesource.com/177723
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>