MALI: platform/chromebook: Init exynos5420 dvfs_infotbl at boot
For a given SoC, the ASV group never changes, and therefore the asv
adjustment to the DVFS voltage table need only happen once.

For exynos5250, there is a sysfs entry to allow turning off the asv
DVFS adjustment.  Thus, there is a possibility that the asv adjustment
may need to be re-applied at runtime.

For exynos5420, asv is aways enabled. Thus, the DVFS voltages never
change so there is no reason to reinitialize the tables on every
DVFS tick.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:24032
TEST=peach_pit: Use ftrace to confirm mali_dvfs_event_proc no longer calls
   get_match_volt 8 times every time.
TEST=cat /sys/devices/11800000.mali/asv
  Confirm frequency:voltage table is correct for SoC's ASV group and number
  of GPU cores (mp4 or mp6).
  => For example:
 asv group:4 mp4
 600000000:975000
 533000000:937500
 480000000:900000
 420000000:887500
 350000000:850000
 266000000:825000
 177000000:825000
 100000000:825000
  Note: The asv tables are in:
       arch/arm/mach-exynos/include/mach/asv-exynos5420.h

Change-Id: I92c06ea345691f5703fbac44265e1d2d14ee4c12
Reviewed-on: https://chromium-review.googlesource.com/177143
Reviewed-by: Abhinav Kochhar <abhinav.k@samsung.com>
Reviewed-by: Akshu Agrawal <akshu@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>