CHROMIUM: mm, oom: Fix race when selecting process to kill
The selection of the process to be killed happens in two spots:
first in select_bad_process and then a further refinement by
looking for child processes in oom_kill_process. Since this is
a two step process, it is possible that the process selected by
select_bad_process may get a SIGKILL just before oom_kill_process
executes. If this were to happen, __unhash_process deletes this
process from the thread_group list. This results in oom_kill_process
getting stuck in an infinite loop when traversing the thread_group
list of the selected process.

Fix this race by adding a pid_alive check for the selected process
with tasklist_lock held in oom_kill_process.

(The upstream discussion over this patch seems stalled at the moment.
Picking it up as a "CHROMIUM" patch for now. Upstream discussion is
here: https://lkml.org/lkml/2013/11/5/471.)

BUG=chromium:256326
TEST=Run the recipe in c#14 in the bug above.

Change-Id: I8f01837b9f95ef3ad0a885c1f8d8c5eddc3cc049
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/177633
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>