CHROMIUM: s5p-mfc: Copy timestamps only when a frame is produced.
Timestamps for destination buffers are assigned by copying them from
corresponding source buffers when the decode operation results in a frame
being outputted to a destination buffer. But the decision when to do this, i.e.
whether the decode operation on current source buffer produced a destination
frame, is wrongly based on "display status". Display status reflects the status
of the destination buffer, not source.

This used to work for firmwares version <= 6, because in addition to the above,
we'd check the decoded frame type register, which was set to "skipped" if
a destination frame was not produced, exiting early from
s5p_mfc_handle_frame_copy_time().
Firmware >=7 does not set the frame type register for frames that were not
decoded anymore though, which results in us wrongly overwriting timestamps of
previously decoded buffers (firmware reports the same destination buffer address
as previously decoded one if a frame wasn't decoded during current operation).

To do it properly, we should be basing our decision to copy the timestamp on the
status of the source buffer, i.e. "decode status". The decode status register
values are confusing, because in its case "display" means "a frame has been
outputted to a destination buffer". We should copy if "decode and display"
is returned in it. There is also no need to check for frame skipped anymore,
also on <= v6 firmware, which behaves in the same way with regards to decode
status register.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=chrome-os-partner:23615,chrome-os-partner:23368
TEST=video playback on v7 and v6 firmware

Change-Id: Ieeb97b4bb0015394121be7301e25f5aaed7b89f2
Reviewed-on: https://chromium-review.googlesource.com/177738
Reviewed-by: Kiran AVND <kiran@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>