BACKPORT: [media] s5p-jpeg: Fix erroneous condition while validating bytesperline value
The aim of the condition is ensuring that the bytesperline
value set by the user space application is proper for the
given format and adjusting it if isn't. As the depth value
of the format description entry is expressed in bits then
the bytesperline value needs to be divided, not multiplied,
by that value to get the number of bytes required to store
single line of image samples.

BUG=chrome-os-partner:33364
TEST=With all the patches in the series run the decoder JPEG test
	./test-jpeg -m 1 -f [input jpeg file] -o [output file]
		-v [decoder node]  -r [format] -s 1
	where:
		format should be a value from 0 to 9

Signed-off-by: Jacek Anaszewski <j.anaszewski@samsung.com>
Signed-off-by: Kyungmin Park <kyungmin.park@samsung.com>
Acked-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>
Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>
(cherry picked from commit cc690904247fdbb169e2544cbbbd4ebd2c05f8cb)
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>

Change-Id: I8ef2cb3f7980b256c2127595dc189945422f6f59
Reviewed-on: https://chromium-review.googlesource.com/226202
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Commit-Queue: Abhilash Kesavan <a.kesavan@samsung.com>
Tested-by: Abhilash Kesavan <a.kesavan@samsung.com>