watchdog: s3c2410_wdt: Only register for cpufreq on CPU_FREQ_S3C24XX
On modern SoCs the watchdog timer is parented on a clock that doesn't
change every time we have a cpufreq change.  That means we don't need
to constantly adjust the watchdog timer, so avoid registering for and
dealing with cpufreq transitions unless we've actually got
CPU_FREQ_S3C24XX defined.

Note that this is more than just an optimization.  The s3c2410
watchdog driver actually pats the watchdog on every CPU frequency
change.  On modern systems these happen many times per second (even in
a system where "nothing" is happening).  That effectively makes any
userspace watchdog program useless (the watchdog is constantly patted
by the kernel).  If we need CPU_FREQ_S3C24XX defined on a
multiplatform kernel we'll need to make sure that kernel supports
common clock and change this to user common clock framework.

This is picked from patchwork at:
  https://patchwork.kernel.org/patch/3235171/

BUG=chrome-os-partner:24222
TEST=With series /dev/watchdog works and reports watchdog reboots

Change-Id: I77ae77f28996730717eb66295bfeb78267831859
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Tomasz Figa <t.figa@samsung.com>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Reviewed-on: https://chromium-review.googlesource.com/177934
Reviewed-by: Olof Johansson <olofj@chromium.org>