CHROMIUM: drm/exynos: Clean up DRM_DEBUG messages
The DRM_DEBUG messages in this driver do a pretty good job of generating
a function trace in dmesg when enabled.

This patch attempts to make this trace more useful in a few ways:
 (1) Instead of dumping file names (which can be trivially deduced from
function names using grep), prefer dumping function parameters, especially
when they are drm object handles.

 (2) Since DRM_DEBUG_* already dump function name; no need to explicitly
print it.

 (3) Use the same style for naming CRTC, CONNECTOR and ENCODER in debug
messages as is used in the drm common code.

 (4) Log when drm objects are created, and include the new object's handle.

 (5) Add debug messages to a few more non-critical path functions.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:218757
TEST=Enable DRM_DEBUG_KMS, reboot, and inspect drm messages in dmesg

Change-Id: Idfbc03c8fe78143f4897f16a760954d075c3c3c8
Reviewed-on: https://chromium-review.googlesource.com/174291
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>