Revert "MALI: Separate PM callbacks for suspend/resume"
This reverts commit 28edd5b2e5ef6c386375f19f62c561eaf06cde3c.

This patch is included in the wk45_2013 release. Due to changes in the
power manager code bettween the wk20_2013 release and later releases
this patch does not apply cleanly. Therefore reverting it, on the
assumption that the wk29 and wk37 releases will not land (without the
wk45_2013 release landing).

Signed-off-by: David Garbett <david.garbett@arm.com>

BUG=chrome-os-partner:28099
TEST=Tested as part of w45_2013 release;
     emerge-daisy chromeos-kernel-next.

Change-Id: I1c72597704a7c0f63129f4b762131237fd87b1ca
Reviewed-on: https://chromium-review.googlesource.com/180951
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>