CHROMIUM: pmcheck: register ramoops volatiles
We had missed to register the pstore console as volatile. As a result, if
there is console output between the pmcheck suspend and resume operations
(say, if there's console output due to CRC recoveries), we end up in an
irrecoverable state and will do a panic reboot.

So, register the backing store as volatile.

There's also a buffer in the pstore ram code that needs similar
registration.

BUG=chromeos-partner:24343
TEST=power_SuspendStress

Change-Id: Ia7fd4532e63d7023afe891ed056c805db97c4be3
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179222
Reviewed-by: Doug Anderson <dianders@chromium.org>