watchdog: s3c2410_wdt: Report when the watchdog reset the system
A good watchdog driver is supposed to report when it was responsible
for resetting the system.  Implement this for the s3c2410, at least on
exynos5250 and exynos5420 where we already have a pointer to the PMU
registers to read the information.

Note that exynos4 SoCs also provide the reset status, but providing
that is left as an exercise for future changes and is not plumbed up
in this patch series.  Also note the exynos4 SoCs don't appear to need
any PMU config, which is why this patch separates the concepts of
having PMU Registers vs. needing PMU Config.

This is picked from patchwork at:
  https://patchwork.kernel.org/patch/3299701/

BUG=chrome-os-partner:24222
TEST=Manual:
1. stop daisydog
2. cat /dev/watchdog0                              # Ignore errors
3. for i in $(seq 100); do echo $i; sleep 1; done  # See 32 prints
4. Wait while device reboots.
5. stop daisydog
6. Run daisydog manually and see:
   /dev/watchdog reported boot status: 0x20 watchdog-timeout
7. reboot
8. Run daisydog manually and see:
   /dev/watchdog reported boot status: normal-boot

Change-Id: Ic5d031eddb1a3bc3fe1f085e364dd499a97a8db0
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Reviewed-on: https://chromium-review.googlesource.com/177935
Reviewed-by: Olof Johansson <olofj@chromium.org>