CHROMIUM: s5p-mfc: Don't allocate codec buffers on STREAMON.
Currently, we allocate private codec buffers on STREAMON, which may fail
if we are out of memory. We don't check for failure though, which will
make us crash with the codec accessing random memory.

We shouldn't be failing STREAMON with out of memory errors though. So move
the allocation of private codec buffers to REQBUFS for OUTPUT queue. Also,
move MFC instance opening and closing to REQBUFS as well, as it's tied to
allocation and deallocation of private codec buffers.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=chrome-os-partner:22388
TEST=unittests, MSE, video playback

Change-Id: I582549e5c33ee105fd9d46042ceed76c3d6e1b07
Reviewed-on: https://chromium-review.googlesource.com/178745
Reviewed-by: Wu-Cheng Li <wuchengli@chromium.org>
Reviewed-by: Arun Mankuzhi <arun@chromium.org>
Tested-by: Arun Mankuzhi <arun@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>